move dashboard to client side so delete button works, remove isDashboard in assessment-table.jsx for not  showing the edit/delete button
